Publishing the services Rest

+++ src/main/java/com/marcelo/xyinc/repository/PointRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import com.marcelo.xyinc.model.Point;
 
-public interface PointRepository extends GenericRepository<Point, Long> {
+public interface PointRepository extends GenericRepository<Point, Integer> {
 
     public List<Point> searchForNearByPoints(final Float coordination_x, final Float coordination_y,
 	    Float maximumDistance_D);
